requirements.txt
amqp==5.1.1
appdata==2.1.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.3.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colored==1.4.3
coverage==6.4.1
Deprecated==1.2.13
Django==4.0.4
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-quill-editor==0.1.40
django-timezone-field==5.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.14.0
feedparser==6.0.10
gevent==21.12.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
kombu==5.2.4
lxml==4.9.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
to-requirements.txt==1.2.0
tomli==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
